Broke down calendar component into seperate files and exported it back into main App to help with individually testing components.
