+++ package.json
@@ -11,7 +11,10 @@
     "lint": "eslint ."
   },
   "jest": {
-    "preset": "react-native"
+    "preset": "react-native",
+    "transformIgnorePatterns": [
+      "node_modules/(?!(react-native|@react-native|react-native-element-dropdown)/)"
+    ]
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "1.21.0",
